Cleaned up and removed the old Ruptures-like 'Continuous Linear Cost' that did not minimize the squared residuals between the trend line and datapoints.
